move calendar props to one file
