--- frontend/src/app/interfaces/AvailabilityCalendarProps.ts
@@ -1,9 +1,0 @@
-export interface AvailabilityCalendarProps {
-  listings: AvailabilityListing[];
-}
-
-export interface AvailabilityListing {
-  id: number;
-  name: string;
-  price: number;
-}
