--- frontend/src/app/interfaces/AvailabilitySidebarProps.ts
@@ -1,8 +1,0 @@
-import { type AvailabilityListing } from "./AvailabilityCalendarProps";
-
-export interface AvailabilitySidebarProps {
-  selectedDates: Date[];
-  blockedDates: Date[];
-  selectedListing: AvailabilityListing;
-  onTabChange: (changed: boolean) => void;
-}
