--- frontend/src/app/interfaces/CalendarDateButtonProps.ts
@@ -1,7 +1,0 @@
-export interface CalendarDateButtonProps {
-  date: Date;
-  isCurrentMonth: boolean;
-  isSelectedDate: boolean;
-  isBlockedDate: boolean;
-  onDateClick: (date: Date) => void;
-}
