--- frontend/src/app/interfaces/ListingDropdownProps.ts
@@ -1,7 +1,0 @@
-import { type AvailabilityListing } from "./AvailabilityCalendarProps";
-
-export interface ListingDropdownProps {
-  listings: AvailabilityListing[];
-  selectedListing: AvailabilityListing;
-  onSelect: (listing: AvailabilityListing) => void;
-}
